Now use enum for viewtype on QML level

+++ AndroidRemoteControl/RemoteImage.cpp
@@ -40,7 +40,7 @@ void RemoteImage::setImageId(int imageId)
         m_imageId = imageId;
         setLabel(ImageStore::instance().label(m_imageId));
         emit imageIdChanged();
-        if (m_type == (int) ViewType::Thumbnails || m_type == (int) ViewType::CategoryItems) { // FIXME, should be two different sizes
+        if (m_type == ViewType::Thumbnails || m_type == ViewType::CategoryItems) { // FIXME, should be two different sizes
             const int size = Settings::instance().thumbnailSize();
             setSize(QSize(size,size));
         }
